Move SGD2MAPI extension, replace use with C89 port
