+++ SGD2FreeResolution/CMakeLists.txt
@@ -517,6 +517,9 @@ set(SRC_PATCHES_REQUIRED_PATCHES
 )
 
 set(SRC_SGD2MAPI_EXTENSION
+    "src/sgd2mapi_extension.h"
+    "src/sgd2mapi_extension.hpp"
+
     "src/sgd2mapi_extension/ddraw_library.c"
     "src/sgd2mapi_extension/ddraw_library_cpp98.cpp"
     "src/sgd2mapi_extension/ddraw_library_version.c"
@@ -537,7 +540,6 @@ set(SRC_SGD2MAPI_EXTENSION
     "src/sgd2mapi_extension/glide3x_library.hpp"
     "src/sgd2mapi_extension/glide3x_library_version.h"
     "src/sgd2mapi_extension/glide3x_library_version.hpp"
-    "src/sgd2mapi_extension/sgd2mapi_extension.hpp"
 
     # DDraw library version
     "src/sgd2mapi_extension/ddraw_library_version/ddraw_library_version_product_name.c"
