+++ SGD2FreeResolution/src/patches/required/glide3x_gr_sst_win_open_patch/glide3x_gr_sst_win_open_patch_sven_1_4_4_21.cc
@@ -47,7 +47,7 @@
 
 #include <stddef.h>
 
-#include "../../../sgd2mapi_extension/sgd2mapi_extension.hpp"
+#include "../../../sgd2mapi_extension.h"
 
 extern "C" {
 
@@ -73,24 +73,24 @@ GrSstWinOpenPatch_Sven_1_4_4_21::GrSstWinOpenPatch_Sven_1_4_4_21()
 
 PatchAddressAndSize
 GrSstWinOpenPatch_Sven_1_4_4_21::GetPatchAddressAndSize01() {
-  ::d2::Glide3xLibraryVersion running_glide3x_library_version =
-      ::d2::glide3x_library_version::GetRunning();
+  D2_Glide3xLibraryVersion running_glide3x_library_version =
+      D2_Glide3xLibraryVersion_GetRunning();
 
   switch (running_glide3x_library_version) {
-    case ::d2::glide3x_library_version::kSven1_4_4_21: {
+    case D2_Glide3xLibraryVersion_kSven1_4_4_21: {
       return PatchAddressAndSize(
           ::mapi::GameAddress::FromOffset(
-              L"glide3x.dll",
+              D2_Glide3xLibrary_GetPath(),
               0xCBA9
           ),
           0xCBB0 - 0xCBA9
       );
     }
 
-    case ::d2::glide3x_library_version::kSven1_4_6_1: {
+    case D2_Glide3xLibraryVersion_kSven1_4_6_1: {
       return PatchAddressAndSize(
           ::mapi::GameAddress::FromOffset(
-              L"glide3x.dll",
+              D2_Glide3xLibrary_GetPath(),
               0xCAD5
           ),
           0xCADC - 0xCAD5
